+++ tvsd/_variables.py
@@ -1,0 +1,6 @@
+from tvsd import state
+
+BASE_PATH = state["base_path"]
+TEMP_BASE_PATH = state["temp_base_path"]
+SERIES_DIR = state["series_dir"]
+SPECIALS_DIR = state["specials_dir"]